Refactor to Game-based architecture with screens
